[Modified] OC-1229 #comment Refactored execution logic, moved operation execution to a separate class #time 1d
